+++ frontend/app/src/entities/proposed-changes/ui/action-button/pc-close-button.tsx
