+++ frontend/app/src/entities/proposed-changes/ui/action-button/pc-reject-button.tsx
